backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic==2.11.9
python-multipart==0.0.20
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.1.1
